Feat: 그룹 삭제시 멤버가 존재하면 삭제할 수 없도록 수정

+++ backend/src/controller/group/deleteGroup.ts
@@ -2,10 +2,11 @@ import { NextFunction, Request, Response } from 'express';
 import { param } from 'express-validator';
 import { validate } from '@validator/middleware/response';
 import { isAdmin } from '@validator/middleware/auth';
-import { selectGroupDetail } from '@service/group/select';
-import { deleteGroup as deleteGroupService } from '@service/group/delete';
 import { getGroupImageDir } from '@uploader/image.uploader';
 import { deleteFile } from '@util/s3';
+import { selectGroupDetail } from '@service/group/select';
+import { deleteGroup as deleteGroupService } from '@service/group/delete';
+import { selectMembersOfGroup } from '@service/member/select';
 
 interface Params {
   groupId: number;
@@ -25,6 +26,9 @@ const controller = async (req: Request, res: Response, next: NextFunction) => {
 
   const [[group]] = await selectGroupDetail(groupId);
   if (!group) return res.status(404).json({ message: '해당 그룹의 데이터가 서버에 존재하지 않아요.' });
+
+  const [members] = await selectMembersOfGroup(groupId);
+  if (members.length > 0) return res.status(400).json({ message: '해당 그룹의 멤버를 먼저 지워주세요.' });
 
   // 기존의 이미지 파일 삭제
   deleteFile({ Key: getGroupImageDir(group.imageName) });
